create more sophisticated message types
