+++ python_reinforcement_learning/python_reinforcement_learning/udp_communicator/udp_receiver.py
@@ -1,5 +1,6 @@
 import asyncio
 import logging
+from .messages import InfoMessage
 
 HOST = "0.0.0.0"
 PORT = 1337
@@ -11,7 +12,7 @@ logging.basicConfig(level=logging.DEBUG)
 logger = logging.getLogger(__name__)
 
 
-class LogRecvProtocol(asyncio.DatagramProtocol):
+class InfoLogRecvProtocol(asyncio.DatagramProtocol):
     def __init__(self):
         super().__init__()
 
@@ -21,12 +22,16 @@ class LogRecvProtocol(asyncio.DatagramProtocol):
 
     # Main entrypoint for processing message
     def datagram_received(self, data, addr) -> None:
-        logger.debug(f"[x] Received Syslog message: {data}")
+        message = InfoMessage.parse(data)
+        if message is not None:
+            logger.debug(f"[x] Received InfoMessage message: {message.info}")
+        else:
+            logger.debug(f"[x] Received corrupt message, expected InfoMessage: {data}")
 
 
 def main():
     logger.info(f"[x] Listening on {HOST}:{PORT}")
-    t = loop.create_datagram_endpoint(LogRecvProtocol, local_addr=(HOST, PORT))
+    t = loop.create_datagram_endpoint(InfoLogRecvProtocol, local_addr=(HOST, PORT))
     loop.run_until_complete(t)  # Server starts listening
     loop.run_forever()
